migrate pvals from teststat
